requirements.txt
openai==0.28
beautifulsoup4
python-dotenv
requests